move objecthandling to session package to solve circular dependencies
